Use decimal instead of double and move logic from Utilities.cs into CalculationService

+++ TaxCalculatorApi/Controllers/CalculateTaxController.cs
@@ -22,7 +22,7 @@
         }
 
         [HttpGet]
-        public CalculationResultDto Get(double totalPackage, int payFrequency)
+        public CalculationResultDto Get(decimal totalPackage, int payFrequency)
         {
             var result = calculationService.CalculateTax(totalPackage, payFrequency);
             return result;
