+++ TaxCalculatorUI/Models/ResultModel.cs
@@ -2,26 +2,26 @@
 {
     public class ResultModel
     {
-        public double TotalPackage { get; set; }
+        public decimal TotalPackage { get; set; }
 
-        public double PayFrequency { get; set; }
+        public decimal PayFrequency { get; set; }
 
-        public double Superannuation { get; set; }
+        public decimal Superannuation { get; set; }
 
-        public double TaxableIncome { get; set; }
+        public decimal TaxableIncome { get; set; }
 
-        public double DeductionTaxableIncome { get; set; }
+        public decimal DeductionTaxableIncome { get; set; }
 
-        public double MedicareLevy { get; set; }
+        public decimal MedicareLevy { get; set; }
 
-        public double BudgetRepairLevy { get; set; }
+        public decimal BudgetRepairLevy { get; set; }
 
-        public double IncomeTax { get; set; }
+        public decimal IncomeTax { get; set; }
 
-        public double Deductions { get; set; }
+        public decimal Deductions { get; set; }
 
-        public double NetIncome { get; set; }
+        public decimal NetIncome { get; set; }
 
-        public double PayPacket { get; set; }
+        public decimal PayPacket { get; set; }
     }
 }
